Renamed and optimized `InterpolatedString`
